Separate out individual and group signups
