+++ package.json
@@ -16,6 +16,7 @@
     "axios": ">=0.21.1",
     "bootstrap": "^4.5.2",
     "bootstrap-icons": "^1.5.0",
+    "clsx": "^1.2.1",
     "firebase": "^9.6.11",
     "history": "^4.10.1",
     "http-proxy-middleware": "^0.19.1",
