+++ src/components/SignUp/Individual/ActorInfo2.tsx
@@ -5,11 +5,10 @@ import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Image from 'react-bootstrap/Image';
 import Form from 'react-bootstrap/Form';
-import { Title } from '../layout/Titles';
-import Checkbox from '../../genericComponents/Checkbox';
-import PrivateLabel from '../../genericComponents/PrivateLabel';
-import { colors, fonts } from '../../theme/styleVars';
-import yellow_blob from '../../images/yellow_blob_2.svg';
+import yellow_blob from '../../../images/yellow_blob_2.svg';
+import { Checkbox, PrivateLabel } from '../../../genericComponents';
+import { colors, fonts } from '../../../theme/styleVars';
+import { Title } from '../../layout';
 
 const ageRanges = [
   '18-22',
