+++ src/components/SignUp/Individual/Credits.tsx
@@ -4,9 +4,9 @@ import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Form from 'react-bootstrap/Form';
-import { Tagline, Title } from '../layout/Titles';
-import InputField from '../../genericComponents/Input';
-import Button from '../../genericComponents/Button';
+import { Tagline, Title } from '../../layout/Titles';
+import InputField from '../../../genericComponents/Input';
+import Button from '../../../genericComponents/Button';
 import 'react-datepicker/dist/react-datepicker.css';
 
 const Credits: React.FC<{
