+++ src/components/SignUp/Individual/Demographics.tsx
@@ -4,11 +4,11 @@ import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import styled from 'styled-components';
 import Image from 'react-bootstrap/Image';
-import InputField from '../../genericComponents/Input';
+import InputField from '../../../genericComponents/Input';
 import Form from 'react-bootstrap/Form';
-import { Title } from '../layout/Titles';
-import { Tagline, TitleThree } from '../layout/Titles';
-import yellow_blob from '../../images/yellow_blob_2.svg';
+import { Title } from '../../layout/Titles';
+import { Tagline, TitleThree } from '../../layout/Titles';
+import yellow_blob from '../../../images/yellow_blob_2.svg';
 
 const websiteTypeOptions = [
   'Personal/Portfolio',
