+++ src/components/SignUp/Individual/OffstageRoles.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import styled from 'styled-components';
 import Accordion from 'react-bootstrap/Accordion';
-import { fonts } from '../../theme/styleVars';
+import { fonts } from '../../../theme/styleVars';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Image from 'react-bootstrap/Image';
 import Container from 'react-bootstrap/Container';
-import { Tagline, Title } from '../layout/Titles';
-import GenericAccordion from '../../genericComponents/GenericAccordion';
-import Checkbox from '../../genericComponents/Checkbox';
-import yellow_blob_1 from '../../images/yellow_blob_1.svg';
+import { Tagline, Title } from '../../layout/Titles';
+import GenericAccordion from '../../../genericComponents/GenericAccordion';
+import Checkbox from '../../../genericComponents/Checkbox';
+import yellow_blob_1 from '../../../images/yellow_blob_1.svg';
 
 const OffstageRoles: React.FC<{
   setForm: any;
